# docs/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Media Viewer API",
        "description": "REST API for Media Viewer - A self-hosted media library browser with WebAuthn/Passkey authentication support",
        "version": "1.0.0",
        "contact": {
            "name": "Media Viewer",
            "url": "https://github.com/djryanj/media-viewer"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/djryanj/media-viewer/blob/main/LICENSE"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development server"
        }
    ],
...
    "paths": {
...
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "List favorite files",
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "favorites": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MediaFile"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add a favorite",
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "path"
                                ],
                                "properties": {
                                    "path": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Added to favorites"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove a favorite",
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "path"
                                ],
                                "properties": {
                                    "path": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Removed from favorites"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "session"
            }
        },
        "schemas": {
...
            "MediaFile": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "image",
                            "video",
                            "folder",
                            "playlist"
                        ]
                    },
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "modifiedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "thumbnailPath": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "isFavorite": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}